output: Move file review model builder
The file-review output package separates display ID mapping, action
selection, page parsing, layout sizing, passive model records, and
persisted state assembly from the terminal renderer.

The renderer still owns change grouping and page model construction. That
forces show commands and file-review list rendering to import the terminal
output module for a data model they need before printing.

This commit addresses that by moving build_file_review_model and its
supporting grouping helpers into output.file_review_model_builder, then
updating source callers to import the builder directly.

Subsequent commits will update command tests and architecture guards to pin
the model-builder boundary.
