+++ src/git_stage_batch/commands/show_from.py
@@ -51,9 +51,9 @@ from ..output.patch import (
     print_gitlink_change,
 )
 from ..output.file_review import (
-    build_file_review_model,
     print_file_review,
 )
+from ..output.file_review_model_builder import build_file_review_model
 from ..data.file_review.pages import normalize_page_spec
 from ..output.file_review_state_builder import (
     make_file_review_state,
